+++ checks/evaluation/maintained_test.go
@@ -16,12 +16,13 @@ package evaluation
 import (
 	"testing"
 
+	"github.com/ossf/scorecard/v4/checker"
 	sce "github.com/ossf/scorecard/v4/errors"
 	"github.com/ossf/scorecard/v4/finding"
 	"github.com/ossf/scorecard/v4/probes/archived"
+	"github.com/ossf/scorecard/v4/probes/createdRecently"
 	"github.com/ossf/scorecard/v4/probes/hasRecentCommits"
 	"github.com/ossf/scorecard/v4/probes/issueActivityByProjectMember"
-	"github.com/ossf/scorecard/v4/probes/notCreatedRecently"
 	scut "github.com/ossf/scorecard/v4/utests"
 )
 
@@ -51,8 +52,8 @@ func TestMaintained(t *testing.T) {
 					Probe:   archived.Probe,
 					Outcome: finding.OutcomeFalse,
 				}, {
-					Probe:   notCreatedRecently.Probe,
-					Outcome: finding.OutcomeTrue,
+					Probe:   createdRecently.Probe,
+					Outcome: finding.OutcomeFalse,
 				},
 			},
 			result: scut.TestReturn{
@@ -72,8 +73,8 @@ func TestMaintained(t *testing.T) {
 					Probe:   archived.Probe,
 					Outcome: finding.OutcomeFalse,
 				}, {
-					Probe:   notCreatedRecently.Probe,
-					Outcome: finding.OutcomeTrue,
+					Probe:   createdRecently.Probe,
+					Outcome: finding.OutcomeFalse,
 				},
 			},
 			result: scut.TestReturn{
@@ -93,8 +94,8 @@ func TestMaintained(t *testing.T) {
 					Probe:   "archvied", /*misspelling*/
 					Outcome: finding.OutcomeTrue,
 				}, {
-					Probe:   notCreatedRecently.Probe,
-					Outcome: finding.OutcomeTrue,
+					Probe:   createdRecently.Probe,
+					Outcome: finding.OutcomeFalse,
 				},
 			},
 			result: scut.TestReturn{
@@ -115,12 +116,40 @@ func TestMaintained(t *testing.T) {
 					Probe:   archived.Probe,
 					Outcome: finding.OutcomeTrue,
 				}, {
-					Probe:   notCreatedRecently.Probe,
-					Outcome: finding.OutcomeTrue,
+					Probe:   createdRecently.Probe,
+					Outcome: finding.OutcomeFalse,
 				},
 			},
 			result: scut.TestReturn{
 				Score:        0,
+				NumberOfWarn: 1,
+			},
+		},
+		{
+			name: "recently created projects get min score",
+			findings: []finding.Finding{
+				{
+					Probe:   hasRecentCommits.Probe,
+					Outcome: finding.OutcomeTrue,
+					Values: map[string]string{
+						hasRecentCommits.NumCommitsKey: "20",
+					},
+				}, {
+					Probe:   issueActivityByProjectMember.Probe,
+					Outcome: finding.OutcomeTrue,
+					Values: map[string]string{
+						issueActivityByProjectMember.NumIssuesKey: "10",
+					},
+				}, {
+					Probe:   archived.Probe,
+					Outcome: finding.OutcomeFalse,
+				}, {
+					Probe:   createdRecently.Probe,
+					Outcome: finding.OutcomeTrue,
+				},
+			},
+			result: scut.TestReturn{
+				Score:        checker.MinResultScore,
 				NumberOfWarn: 1,
 			},
 		},
