+++ probes/createdRecently/def.yml
@@ -12,17 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-id: notCreatedRecently
-short: Checks that the project was not created in the last 90 days.
+id: createdRecently
+short: Checks if the project was created in the last 90 days.
 motivation: >
   When Scorecard checks the activity of a project in the last 90 days, the project may not have been created before the last 90 days. As such, Scorecard cannot give an accurate score. This probe helps Scorecard assess whether it can give an accurate score when checking the project activity in the last 90 days.
 implementation: >
   The implementation checks the creation date is within the last 90 days.
 outcome:
-  - If the project was created within the last 90 days, the outcome is OutcomeFalse (0).
-  - If the project was created before the last 90 days, the outcome is OutcomeTrue (1). The finding will include a "lookBackDays" value which is the time period that the probe looks back in. 
+  - If the project was created within the last 90 days, the outcome is OutcomeTrue.
+  - If the project is older than 90 days, the outcome is OutcomeFalse. The finding will include a "lookBackDays" value which is the time period that the probe looks back in. 
 remediation:
-  onOutcome: False
+  onOutcome: True
   effort: Low
   text:
     - The only remediation for this probe is to wait until 90 days have passed after a project has been created.
