+++ probes/createdRecently/impl_test.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 //nolint:stylecheck
-package notCreatedRecently
+package createdRecently
 
 import (
 	"testing"
@@ -44,7 +44,7 @@ func Test_Run(t *testing.T) {
 				},
 			},
 			outcomes: []finding.Outcome{
-				finding.OutcomeFalse,
+				finding.OutcomeTrue,
 			},
 		},
 		{
@@ -55,7 +55,7 @@ func Test_Run(t *testing.T) {
 				},
 			},
 			outcomes: []finding.Outcome{
-				finding.OutcomeTrue,
+				finding.OutcomeFalse,
 			},
 		},
 	}
